+++ src/evaln/sub_and_add.c
@@ -12,25 +12,25 @@
 
 #include "eval_expr.h"
 
-void	addos(int *stackos, int c, t_int *lastint)
+void	addos(long long *stackos, int c, t_int *lastint)
 {
 	stackos[lastint->ol] = c;
 	(lastint->ol)++;
 }
 
-void	subos(int *stackos, t_int *lastint)
+void	subos(long long *stackos, t_int *lastint)
 {
 	stackos[lastint->ol - 1] = 0;
 	(lastint->ol)--;
 }
 
-void	addzn(int *stackzn, t_calc_tkntype c, t_int *lastint)
+void	addzn(long long *stackzn, t_calc_tkntype c, t_int *lastint)
 {
 	stackzn[lastint->zl] = c;
 	(lastint->zl)++;
 }
 
-void	subzn(int *stackzn, t_int *lastint)
+void	subzn(long long *stackzn, t_int *lastint)
 {
 	stackzn[lastint->zl - 1] = 0;
 	(lastint->zl)--;
